+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 CHANGELOG
 ==========================
 
+## 1.1.0 Alpha 2 (`1010012`)
+
+- **New:** Seed for media gallery comment (#12)
+- **New:** Seed for media gallery item reaction (#13)
+- **New:** Seed for media gallery comment reaction (#14)
+- **New:** Seed for media gallery album reaction (#15)
+- **New:** Seed for profile post (#16)
+- **New:** Seed for profile post comment (#17)
+- **New:** Seed for profile post reaction (#18)
+- **New:** Seed for profile post comment reaction (#19)
+
 ## 1.1.0 Alpha 1 (`1010011`)
 
 - **New:** Seed for media gallery category (#7)